requirements.txt
discord.py
python-dotenv
httpx
aiosqlite
aiohttp
